Add product test and refactor erros attributes

+++ src/main/java/br/com/restassured/automation/model/request/product/AddOrUpdateProductRequest.java
@@ -15,7 +15,7 @@ public class AddOrUpdateProductRequest {
 
     private String descricao;
 
-    private BigDecimal preco;
+    private int preco;
 
     private int quantidade;
 }
